[http] Refactor common HTTP request properties into a superclass
